h753/Lib/mavlink_hal/config/requirements.txt
# Python dependencies for MAVLink HAL configuration tools

# Schema validation
jsonschema>=4.17.0
pyyaml>=6.0

# Optional: Enhanced validation and reporting
colorama>=0.4.6  # Colored terminal output
tabulate>=0.9.0  # Pretty table formatting

# Optional: Code generation (for Epic2 Task2)
jinja2>=3.1.2    # Template engine
